+++ lib/src/utilities/barrel_export_utilities.dart
@@ -1,4 +1,3 @@
 export 'converter.dart';
-export 'multiple_getter_persistent_repository.dart';
 export 'cornerstone_snapshot.dart';
 export 'hive_persistent_repository_mixin.dart';
